# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales Support API",
        "description": "SES企業向け営業支援システム API",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "ローカル開発環境"
        }
    ],
    "paths": {
...
        "/api/v1/cards": {
            "get": {
                "tags": [
                    "BusinessCards"
                ],
                "summary": "名刺一覧取得",
                "operationId": "8cf23f8988ac1a3393482225444acf75",
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "BusinessCards"
                ],
                "summary": "名刺画像OCR解析・登録（Google Vision + Claude API）",
                "operationId": "a769d795ae4ba73b0f25c5b2552b81f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "images[]": {
                                        "description": "JPEG/PNG画像（最大20枚・各10MB）",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "登録成功（顧客・担当者も自動登録）"
                    },
                    "422": {
                        "description": "バリデーションエラー"
                    },
                    "500": {
                        "description": "OCR処理失敗"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cards/{id}": {
            "get": {
                "tags": [
                    "BusinessCards"
                ],
                "summary": "名刺詳細取得",
                "operationId": "285a648c51ca1005f19ba78a0363fa3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "名刺ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "404": {
                        "description": "見つかりません"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "BusinessCards"
                ],
                "summary": "名刺情報更新",
                "operationId": "123a17dfccefda477d296bca3ca5010f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "名刺ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "422": {
                        "description": "バリデーションエラー"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "BusinessCards"
                ],
                "summary": "名刺削除（Supabase Storageからも削除）",
                "operationId": "46554027ad9db59be9018fc18e6214a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "名刺ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "削除成功"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Supabase Auth JWT トークン",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}